game: Calculate the visible area by unprojecting the corners of vp
The previous method for calculating the visible area relied on the
camera looking directly overhead the centre of the world. This
approach instead calculates the x/y positions of the intersections
between the lines projected by the corners of the viewport with the
plane of the bottom of the world and then grabs the bounding box of
those points. That way it will also work if the base transform is more
complicated.