Move the data directory under src
If a binary distribution is made then ideally the executable should be
in the root of the zip or tarball and the data directory should be at
the same level. In order to do this without breaking running it from
the src directory then it's easiest to just put the data directory in
the src directory so that the layout matches that of the final
distribution.