Make the map texture have multiple columns
Instead of just a single long vertical column of block images there
can now be multiple columns with horizontal padding in-between. The
main reason to do this is to avoid hitting the maximum texture
dimension limits which are only guaranteed to be at least 1024.

The texture is also now no longer forced to be a power of two because
there is no reason to do that on GL >= 2.0.