Add support for different types of NPC movement
The initial NPC state now contains an enum for the type of movement.
Currently there is only static movement which is effectively what was
already implemented. fv_logic now keeps track of a state for the
player so it can detect when it has returned to its normal position
after being afraid. The static movement now also returns to the
original orientation once its stopped being afraid.