Move the decision to clear to fv-main
There is now a function in fv-game to determine whether painting would
cover the framebuffer given a world center position. That way fv-main
can determine whether to clear. This has the advantage that it can
clear the color and depth buffers at the same time and it seems like a
more sensible place to have it because fv-game conceptually isn't the
only thing that paints. The center value is now also passed down to
fv_game_paint which will also make it easier later to handle
multi-player split-screen.