logic: Add a state
The state marks whether the logic is currently running. This provides
a way to use the fv_logic before the game has started. It will also
be used when the game ends. A new function called fv_logic_reset sets
the state to playing and also sets the number of players. This is
currently just hardcoded to reset the number of players to 1 in
fv-main.