main: Support multiple players, joysticks and mice
The game now starts off displaying the HUD with the title screen and a
chance to choose the number of players. Each player is then given a
chance to configure their direction keys. The screen becomes split if
there is more than 1 player. Most of the code for handling input was
taken from Regular Octagon.