Add a fallback for GL_ARB_instanced_arrays
If the extension is not available then each person and special model
will be drawn with a separate draw call. In that case the instanced
attributes are converted to uniforms. In practice this doesn't seem to
make any measurable performance difference and it means the game can
work on unextended GL 3.1.