Add a fallback if texture 2D arrays are not supported
The code path for not using instanced rendering in fv_person_painter
now uses multiple textures instead of a texture 2D array. If either
instanced rendering or texture 2D arrays are not supported then it
will fall back to this code path. It could be possible to retain the
third code and maybe add a fourth that takes advantage of just one of
either of the extensions but that's probably getting a bit silly.