Don't share buffers with vertices and indices
WebGL doesn't let you bind a buffer to a different target once it has
been bound to another target. That makes it impossible to use a single
buffer for both indices and vertices. I'm not really sure whether this
was a good idea anyway so this patch just makes it always use two
separate buffers.