Enable multisampling
It will now try to create an SDL window with a multisampling context
with at least 2 samples. If this fails it will try again with
multisampling disabled. Multisampling is disabled for rendering the
HUD because it shouldn't make any difference in that case and it seems
to make a measurable performance improvement at least on Intel
hardware.