Load all of the images at once with a callback event
Instead of all of the painting code manually loading images they are
now all loaded upfront using a new fv_image_data object. This is
passed to the constructors for the painters. The fv_image_data pushes
an SDL event when it has finished loading. There are helper functions
to set an image of a texture using a loaded image. The GL context and
painter objects are now only created once all of the images are
loaded. This will make it easier to implement a special empscripten
version that uses the HTML DOM to download and decompress the images.