map-buffer: Reset the buffer data to NULL if the map failed
Part of the nVidia GL recommendations says that you should call
glBufferData with NULL if you are mapping a buffer for write and you
don't need the old contents so that the driver can know that it can
discard it. In order to do this the fv_map_buffer function needs to
take the usage parameter.