map: Encode the normal as the fourth component of the position
The normal is now encoded as the fourth component of the position
rather than its own component because I read somewhere that all
attributes should be aligned to a float. I'm not sure if this is true
or not but it's not really difficult to do so we might as well play it
safe.