Concatenate strings rather than multiple shaders per stage
When multiple sources are used for a single stage with in a program,
instead of creating a separate shader object for each source and
linking them all together, it now just concatenates the sources and
makes a single shader. This is required for WebGL which can't have
multiple shaders per stage.