Don’t overwrite buffers midframe when painting several viewports
Previously the various painters weren’t aware when they were being
painted more than once in a single frame to implement the multiple
viewports. That meant they would just overwrite the data in buffers
that are still pending in the command buffers and weird things would
happen. To fix this all of the painters now have begin and end frame
functions so that they can recognise this situation.