main: Handle resizing from the surface caps instead of ConfigureNotify
The validation layers get upset if we don’t always use the size
matching the last call to vkGetPhysicalDeviceSurfaceCapabilitiesKHR.
vkAcquireNextImageKHR can also return an error code when the size
changes which was previously causing the game to just quit. To fix
this it now exclusively listens to size changes based on this error
and the surface caps.