Increase the size of the instance buffers
On the Mesa i965 driver at least, the buffers get rounded up to a
multiple of 4KB so there’s no point in allocating anything less than
that. This is probably a reasonable value to assume so it now always
allocates the buffers to as close as possible to this size.