window: Don’t recreate resources if SUBOPTIMAL from vkAcquireImage
If VK_SUBOPTIMAL_KHR is returned from vkAcquireImageKHR then the
acquisition is considered a success which means the semaphore will be
submitted. If we then try to reuse it without waiting on it the
validation might report an error because it could be signalled. To fix
this we now just go ahead with the render when SUBOPTIMAL is returned.
This should still end up recreating the resources when SUBOPTIMAL is
returned again from vkQueuePresentKHR.